% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_association_matrices.R
\name{association}
\alias{association}
\alias{nonassociative}
\alias{unstructured}
\alias{uniform}
\title{Association Matrices for Fitting the Multivariate Box-Cox Symmetric Distributions Generated by
    a Normal Scale Mixture Copula}
\usage{
nonassociative(d)

unstructured(d)

uniform(d)
}
\arguments{
\item{d}{dimension of the association matrix.}
}
\value{
An object of class \code{"association"}, mainly used in the \code{bcnsm_fit} function. It
     returns a list with the components
        \describe{
          \item{npar}{number of parameters relative to the association structure.}
          \item{start}{a function \code{function(x)} which returns the initial values for the
              parameters related to the association matrix based on a matrix data \code{x}.}
          \item{Gamma}{a function \code{function(gamma)} that returns the association matrix as a
              function of the parameters \code{gamma}.}
          \item{lower, upper}{lower and upper bounds of the possible values assumed by the
              parameters relative to the association matrix.}
          \item{name}{character, with the name of the association structure.}
        }
}
\description{
Class of association matrices available in the \code{bcnsm} package.
}
\details{
The normal scale mixture copula entirely describes the dependence structure of the
    BCNSM distributions. Because copulas are invariant under strictly increasing transformations,
    copula-based measures of association of the scale mixtures of normal distributions are the
    same for the BCNSM distributions. This statement is valid for measures of concordance, such
    as Kendall's tau or the tail dependence coefficients.

    Let \code{gamma_ij} be the element of the ith row and jth column of the association matrix,
    say a d-dimensional \code{Gamma}, where \code{gamma_ii = 1}, for \code{i, j = 1, ..., d}. If
    \code{Y = (Y1, ..., Yd)}' is a d-dimensional random vector with a BCNSM distribution and with
    association matrix \code{Gamma}. Thus, Kendall's tau of \code{Yi} and \code{Yj} is given by

    \code{tau(Yi, Yj) = 2 * asin(gamma_ij) / pi},

    for \code{i, j = 1, ..., d} and it is invariant under the copula. Since Kendall’s tau is a
    rank correlation coefficient, it measures any monotonous association between two random variables.
    Thus, we have an increasing monotonous association between \code{Yi} and \code{Yj} when \code{gamma_ij}
    is close to 1, and a decreasing monotonous association when \code{gamma_ij} is close to -1.
    Furthermore, \code{gamma_ij = 0} indicates no monotonous association. These facts show that
    the association matrix \code{Gamma} plays an essential role in specifying the dependence
    structure of \code{Y}. It can be parameterized in terms of a real-valued vector \code{gamma}
    in an interpretative way depending on the type of association that is of interest to describe.
    Currently, the \code{bcnsm} package provides the non-associative, unstructured, and
    exchangeable association structures.
}
\examples{
## Non-associative
nonassociative()
nonassociative(4)$Gamma()

## Unstructured
unstructured(4)
unstructured(4)$Gamma(c(0.2, 0.3, 0.4, 0.5, 0.6, 0.2))

## Uniform
uniform()
uniform(4)$Gamma(0.6)

}
\references{
Medeiros, R. M. R. de, and Ferrari, S. L. P. (2024). Multivariate Box-Cox symmetric
     distributions generated by a normal scale mixture copula.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
