% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_association_matrices.R
\name{association_methods}
\alias{association_methods}
\alias{print.association}
\title{Methods for 'association' objects}
\usage{
\method{print}{association}(x, ...)
}
\arguments{
\item{x}{an object of class \code{association}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Methods for 'association' objects
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
