% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_methods.R
\name{bcnsm-methods}
\alias{bcnsm-methods}
\alias{vcov.bcnsm}
\alias{logLik.bcnsm}
\alias{AIC.bcnsm}
\alias{print.bcnsm}
\alias{summary.bcnsm}
\alias{print.summary.bcnsm}
\title{Methods for \code{"bcnsm"} objects}
\usage{
\method{vcov}{bcnsm}(object, par = c("all", "mu", "sigma", "lambda", "nu", "gamma"), ...)

\method{logLik}{bcnsm}(object, ...)

\method{AIC}{bcnsm}(object, ..., k = 2)

\method{print}{bcnsm}(x, digits = 2L, ...)

\method{summary}{bcnsm}(object, digits = 4L, ...)

\method{print}{summary.bcnsm}(x, ...)
}
\arguments{
\item{par}{character; specifies which submatrix of the asymptotic covariance matrix of the
maximum likelihood estimators should be returned. The options are \code{"all"} (default),
\code{"mu"}, \code{"sigma"}, \code{"lambda"}, \code{"nu"}, and \code{"gamma"}.}

\item{...}{further arguments passed to or from other methods.}

\item{k}{numeric, the penalty per parameter to be used; the default
\code{k = 2} is the classical AIC.}

\item{x, object}{an object of class \code{bcnsm}.}

\item{digits}{number of digits in print methods.}
}
\description{
Methods for \code{"bcnsm"} objects
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
