% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_bcnsm_distributions.R
\name{bcnsm}
\alias{bcnsm}
\alias{dbcnsm}
\alias{rbcnsm}
\title{The Class of the Multivariate Box-Cox Symmetric Distributions Generating
    by a Normal Scale Mixture Copula}
\usage{
dbcnsm(
  x,
  mu,
  sigma,
  lambda,
  nu = NULL,
  Gamma = diag(ncol(x)),
  copula = c("gaussian", "t", "slash", "hyp"),
  delta = NULL,
  margins = "bcno",
  log = FALSE
)

rbcnsm(
  n,
  mu,
  sigma,
  lambda,
  nu,
  Gamma = diag(ncol(x)),
  copula = c("gaussian", "t", "slash", "hyp"),
  delta = NULL,
  margins = "bcno"
)
}
\arguments{
\item{x}{vector or matrix of non-negative quantiles. If \code{x} is a matrix,
each row is taken to be a quantile.}

\item{mu}{vector of the marginal scale parameters in \code{(0, infty)^d}, where \code{d} is the
dimension.}

\item{sigma}{vector of the marginal relative dispersion parameters in \code{(0, infty)^d}, where \code{d} is the
dimension.}

\item{lambda}{vector of the marginal skewness parameters in \code{R^d}, where \code{d} is the
dimension.}

\item{nu}{vector of possible extra parameters of the marginal distributions. In the case where
no marginal is indexed by \code{nu}, it must be specified with \code{NA}.}

\item{Gamma}{the association matrix. It must be a positive-definite correlation matrix, default is
\code{diag(ncol(x))} (the so-called non-associative association structure).}

\item{copula}{character; informs which normal scale mixture distribution
should be used to generate the NSM copula. Currently,
the copulas available are: Gaussian (\code{"gaussian"}),
Student's t (\code{"t"}), slash (\code{"slash"}), and hyperbolic (\code{"hyp"}).}

\item{delta}{possible extra parameter associated with the mixing distribution of the
copula. For example, the degrees of freedom of the \code{t} copula, or the heavy tailness
parameter of the slash or the hyperbolic copula.}

\item{margins}{a character or a character vector; specifies the marginal BCS distributions. If all
BCS margins are the same, it is sufficient to enter only one character. A table with the
current available BCS distributions can be seen in \code{\link[bcnsm]{bcs}}.}

\item{log}{logical; if \code{TRUE}, probabilities \code{p} are given
as \code{log(p)}.}

\item{n}{number of replicates to return.}
}
\value{
\code{dbcnsm} returns the evaluated joint density function and
     \code{rbcnsm} generates random values.
}
\description{
These functions provide the joint probability density function
    and a random generator for the class of the multivariate Box-Cox symmetric
    distributions generated by a normal scale mixture copula.
}
\examples{
\dontrun{

### Sample size and dimension
n <- 1000
d <- 4

### Association matrix
Gamma <- matrix(0.8, d, d)
diag(Gamma) <- 1

### Marginal specifications

# Marginals
margins <- c("bchp", "bcsl", "bcpe", "bcloii")

# Marginal parameters
mu <- c(19, 20, 15, 20)
sigma <- c(0.2, 0.6, 0.4, 0.3)
lambda <- c(-1, 1.2, 0, 1.6)
nu <- c(6, 4, 8, NA)

### Copula
copula <- "slash"
delta <- 3

### Generating observations
y <- rbcnsm(n, mu, sigma, lambda, nu, Gamma, copula, delta, margins)

mvplot(y)  ## See ?mvplot for documentation
}

}
\references{
Vanegas, L. H., and Paula, G. A. (2016). Log-symmetric distributions:
     statistical properties and parameter estimation. \emph{Brazilian Journal of
     Probability and Statistics}, 30, 196--220.

 Ferrari, S. L. P., and Fumes, G. (2017). Box-Cox symmetric distributions and
     applications to nutritional data. \emph{AStA Advances in Statistical Analysis},
     101, 321--344.
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
